Rename lcos module to lcos_distributed
