+++ libs/include/include/hpx/include/lcos.hpp
@@ -22,8 +22,8 @@
 #include <hpx/include/actions.hpp>
 #include <hpx/lcos/base_lco.hpp>
 #include <hpx/lcos/base_lco_with_value.hpp>
-#include <hpx/lcos/channel.hpp>
 #include <hpx/lcos/packaged_action.hpp>
+#include <hpx/lcos_distributed/channel.hpp>
 #endif
 
 #endif
