+++ libs/lcos_distributed/README.rst
@@ -6,11 +6,11 @@
     Distributed under the Boost Software License, Version 1.0. (See accompanying
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-====
-lcos
-====
+================
+lcos_distributed
+================
 
 This library is part of HPX.
 
 Documentation can be found `here
-<https://stellar-group.github.io/hpx-docs/latest/html/libs/lcos/docs/index.html>`__.
+<https://stellar-group.github.io/hpx-docs/latest/html/libs/lcos_distributed/docs/index.html>`__.
