+++ libs/lcos_distributed/examples/CMakeLists.txt
