+++ libs/lcos_distributed/include_compatibility/hpx/lcos/server/channel.hpp
@@ -1,0 +1,19 @@
+//  Copyright (c) 2020 ETH Zurich
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/config.hpp>
+#include <hpx/lcos_distributed/config/defines.hpp>
+#include <hpx/lcos_distributed/server/channel.hpp>
+
+#if defined(HPX_LCOS_DISTRIBUTED_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_MSVC)
+#pragma message("The header hpx/lcos/server/channel.hpp is deprecated, \
+    please include hpx/lcos_distributed/server/channel.hpp instead")
+#else
+#warning "The header hpx/lcos/server/channel.hpp is deprecated, \
+    please include hpx/lcos_distributed/server/channel.hpp instead"
+#endif
+#endif
