+++ libs/lcos_distributed/tests/regressions/channel_not_empty_2890.cpp
@@ -10,7 +10,7 @@
 #include <hpx/testing.hpp>
 
 #if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
-#include <hpx/lcos/channel.hpp>
+#include <hpx/lcos_distributed/channel.hpp>
 
 HPX_REGISTER_CHANNEL(int);
 #endif
