+++ libs/lcos_distributed/tests/regressions/channel_register_as_2722.cpp
