+++ libs/lcos_distributed/tests/unit/CMakeLists.txt
@@ -14,7 +14,7 @@ foreach(test ${tests})
 
   source_group("Source Files" FILES ${sources})
 
-  set(folder_name "Tests/Unit/Modules/LCOs")
+  set(folder_name "Tests/Unit/Modules/LCOsDistributed")
 
   add_hpx_executable(${test}_test
     INTERNAL_FLAGS
@@ -24,5 +24,5 @@ foreach(test ${tests})
     HPX_PREFIX ${HPX_BUILD_PREFIX}
     FOLDER ${folder_name})
 
-  add_hpx_unit_test("modules.lcos" ${test} ${${test}_PARAMETERS})
+  add_hpx_unit_test("modules.lcos_distributed" ${test} ${${test}_PARAMETERS})
 endforeach()
